requirements.txt
python_dateutil >= 2.5.3
setuptools >= 65.5.0
urllib3 >= 1.25.3
pydantic >= 1.10.2
aenum >= 3.1.11
